Fix staticcheck warnings
